+++ src/components/section/admin/sideNavbar.jsx
@@ -16,7 +16,7 @@ const sideNavbar = (props) => {
             <NavbarButton endpoint="/admin/dashboard" icon="fas fa-home" label="Dashboard"/>
             <NavbarButton endpoint="/admin/dashboard" icon="fas fa-prescription-bottle-alt" label="Product List"/>
             <NavbarButton endpoint="/admin/dashboard" icon="fas fa-tablets" label="Material List"/>
-            <NavbarButton endpoint="/admin/userHistory" icon="fas fa-user" label="User History"/>
+            <NavbarButton endpoint="/admin/userDatas" icon="fas fa-user" label="User History"/>
             <br/>
             <p style={{ color: "var(--pink-color)", marginBottom: 5, fontSize: 16 }}>STATS</p>
             <NavbarButton endpoint="/admin/sales" icon="fas fa-credit-card" label="Sales Report"/>
